Remove lerna - we will use `link:` specifiers instead
I tried to get lerna working with hoisting, and un-published packages;
but I could not resolve all the issues. The only thing that we really
need mono-repo support for is hoisting certain dependencies - React in
particular fails if multiple copies are loaded. We can work around that
problem by installing common dependencies at the top-level.

For more information on the `link:` specifier see:

https://github.com/npm/npm/pull/15900