poodle-service: redesign composer API
The new API is designed to produce `Message` values instead of
`MessageBuilder`. We introduce a `serialize` module to convert from
`Message` to `MessageBuilder`. This allows an app to store and display
newly-composed messages alongside other `Message` values. The new API is
also an attempt to reduce the code required to write a compose function.